+++ src/test/scala/scorex/lagonaki/integration/api/SeedAPISpecification.scala
@@ -12,6 +12,7 @@ class SeedAPISpecification extends FunSuite with Matchers {
   test("/seed API route") {
     Base58.decode((getRequest("/seed/") \ "seed").as[String]).isSuccess shouldBe true
   }
+
   test("/seed/{length} API route") {
     val length = Random.nextInt(4096)
     Base58.decode((getRequest(s"/seed/$length") \ "seed").as[String]).get.length shouldBe length